+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  - dynamically load the schema for the `jobConfig`
  - adjust scheduling `jobs` to be expandable and used by `plugins`
  - modify column `config` to be `longtext` in table `tasks`
+ - catch `checkbox_value` if sent for a `input` from type `checkbox`
 
 ### Bugfix
 - bugfix the live search in the dropdown for selecting groups for job actions from type `add_group` or `remove_group`;